// src/app/models/CreditorDel_model.ts
export interface DelCreditorObj {
    SessionUserId: string;
    CountryCode: string;
    SessionId: string;
    ViewUser: string;
    LanguageCode: string;
    PaymentCategory: string;
    IbActionTimeStamp: string;
    InputFromScreen: {
        CreditorId: string;
        CreditorStatus: string;
        TemplateId: string;
        TemplateStatus: string;

    }

}